# feat(perf-detector-threshold-configuration): Added extended detector for N+1 API Calls Issue