- Added new detector: NPlusOneDBSpanDetectorExtended
- Detector only collects metrics. Doesn't create issues.
- New detector settings:
<img width="353" alt="Screenshot 2023-07-06 at 1 42 13 PM" src="https://github.com/getsentry/sentry/assets/60121741/13571665-4897-4ed6-a46f-0043cf13ac84">

- Got rid of individual span duration. We should be creating an issue when 1000, 10ms eligible spans potentially leads to 10s performance overhead. Currently we don't.
- Reduced count threshold from 10 to 5. This aligns with N plus one db, count threshold.
-  Increased concurrency threshold from 5ms to 15ms. This was based off this [slack thread](https://sentry.slack.com/archives/C051RMBHG4F/p1686328559140749). 
- Added tests.