@Abdkhan14 2023-07-06
> Logic looks good!
> 
> I don't necessarily want to hold up the PR on this since it's a short-lived change anyway, but as an idea for future changes like this: one way to make extended detectors easier to review (and keep their logic in sync with their parent detector) is to extract the logic that's going to change into its own helper method so you can override just that, instead of having to override big chunks like `visit_span` or `_maybe_store_problem`. ([Here's an example](https://github.com/getsentry/sentry/pull/49805/files#diff-d862954a47b2de0eb3a8ce9baa2d96182b85cb0b1389295c9ecc1b4def8ba734), where I extracted `_is_slower_than_threshold` so that only that bit had to change.)

Will keep this in mind moving forward. Thanks @mjq-sentry 💯